add socket class
